99% Complete Schematic and Footprints.
The only items on the schematic that are not complete are the switches, as the 5th switch needs to be defined and all of them should be renamed.

Some of the Teensy pins are not named properly either. Besides that, every component should be correctly associated with an accurate footprint.