Grounded DRV8871s, sizde traces, and general neatness
There are now vias to for the motor drivers to manage heat. Capacitors were moved slightly. Traces are now sized, and aligned.

There is huge overhead for each trace, with the exception of signal.

The two issues that still need to be addressed are the XT60s and the Relay, both of which are not going to the right pins.